backend/requirements.txt
django==2.2.16
django-cors-headers==3.2.1
djangorestframework==3.11.0
mysqlclient==1.3.12
